move helper functions from handlers.go to helpers.go
